update column name searching for ozone
Only updated column name searching for ozone, will need do the samething for other gas.

This update is due to now we are using Blick v1.5.0 for Pan103/104, so the L2 data has different header!

Note: The Pan108/109/122 data are still processed by Blickv1.4.1. We can't update them to v1.5.0, since the CF files for them are not compatiable with v1.5.0!!!!